deleted unnecessary files an added a short note to About.svelte
