+++ src/routes.js
@@ -1,8 +1,6 @@
 // Components
 import Home from './routes/Home.svelte'
-import Name from './routes/Name.svelte'
 import About from './routes/About.svelte'
-import Wild from './routes/Wild.svelte'
 import NotFound from './routes/NotFound.svelte'
 
 // Export the route definition object
@@ -10,15 +8,8 @@ export default {
     // Exact path
     '/': Home,
 
-    // Using named parameters, with last being optional
-    '/hello/:first/:last?': Name,
-
     // Wildcard parameter
-    // Included twice to match both `/wild` (and nothing after) and `/wild/*` (with anything after)
-    '/wild': Wild,
-    '/wild/*': Wild,
-
-
+    // Included twice to match both `/About` (and nothing after) and `/About/*` (with anything after)   
     '/About': About,
     '/About/*': About,
